+++ lib/dollar.dart
@@ -1,5 +1,4 @@
 library dollar;
 
 export 'src/core.dart';
-export 'src/effects.dart';
-export 'src/expressions.dart';
+export 'src/extensions.dart';